IrisHelper; Minor Updates to README
Starting from this revision, we are also including more thorough parts
of our code that was customized for better usage of CakePHP in our
projects, and handy snippets of code that may serve as useful in your
own CakePHP (and Twitter Bootstrap-based) projects - see the
`View/Helper/IrisHelper.php` or `View/Helper/IrisIOHelper.php` for
example.
We hope that our contributions to the Open-Source community can benefit
everyone.